requirements.txt
addfips==0.3.1
gurobipy==10.0.0
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.1
PyYAML==6.0
scikit_learn==1.2.0
seaborn==0.12.1
xgboost==1.7.1
